chore: move styles to style.ts files
