+++ packages/app-explorer/src/systems/Transaction/component/TxScripts/TxReceiptHeader/styles.ts
@@ -1,0 +1,7 @@
+import { tv } from 'tailwind-variants';
+
+export const styles = tv({
+  slots: {
+    header: 'group min-h-[42px] gap-2 tablet:gap-4',
+  },
+});
