+++ packages/app-explorer/src/systems/Transaction/component/TxScripts/TxScriptsContent/styles.ts
